chore: moves dynclient to client pkg
